fix: restore correct test file test_herald_publisher.py to match workflow requirements
